Extract methods into init mixin
