+++ src/core/index.js
@@ -3,7 +3,18 @@ import { UrlMixin } from './mixins/url';
 import { RequestMixin } from './mixins/request';
 import { CrudMixin } from './mixins/crud';
 import { CustomRoutesMixin } from './mixins/custom-routes';
-import Rapid from './request';
+import { warn } from '../utils/debug';
+
+function Rapid(config) {
+  if (process.env.NODE_ENV !== 'production' &&
+    !(this instanceof Rapid)
+  ) {
+    warn('Rapid is a constructor and should be called with the `new` keyword');
+  }
+
+  this._init(config);
+  this.applyCallableRequestMethods();
+}
 
 InitMixin(Rapid);
 UrlMixin(Rapid);
